Make Git externals optional

+++ vars/retrieveExternals.groovy
@@ -1,18 +1,21 @@
 #!/usr/bin/env groovy
 
 def call() {
+    
+    def exists = fileExists 'git_externals.json'
 
-    echo "Retrieve externals"
+    echo "Retrieve externals ${exists}"
     
-    // TODO no-op if no get_externals.json
-    
-    withCredentials([sshUserPrivateKey(credentialsId: env.GITHUB_CREDENTIAL_ID, keyFileVariable: 'jenkins_private_key')]) {
-        // Want to throw away the noisy output; TODO ${jenkins_private_key} here?
-        sh returnStdout:true, script: '''
-        which ssh-agent || ( apt-get update -y && apt-get install openssh-client -y )
-        eval $(ssh-agent -s)
-        ssh-add ${jenkins_private_key}
-        git externals update
-        '''
+    // Make Git externals optional
+    if (exists) {
+        withCredentials([sshUserPrivateKey(credentialsId: env.GITHUB_CREDENTIAL_ID, keyFileVariable: 'jenkins_private_key')]) {
+            // Want to throw away the noisy output; TODO ${jenkins_private_key} here?
+            sh returnStdout:true, script: '''
+            which ssh-agent || ( apt-get update -y && apt-get install openssh-client -y )
+            eval $(ssh-agent -s)
+            ssh-add ${jenkins_private_key}
+            git externals update
+            '''
+        }
     }
 }
